refactor(app): moves ContactRequest types into own type file
