+++ app/src/api/contacts/acceptOrDeclineContactRequest/acceptOrDeclineContactRequest.ts
@@ -1,5 +1,8 @@
 import { apiClient } from "../../client";
-import { ContactRequest, ContactRequestSchema } from "../../../types/Contact";
+import {
+  ContactRequest,
+  ContactRequestSchema,
+} from "../../../types/ContactRequest";
 
 export const acceptOrDeclineContactRequest = async ({
   contactRequestId,
