+++ app/src/api/contacts/sendContactRequest/sendContactRequest.ts
@@ -1,5 +1,8 @@
 import { apiClient } from "../../client";
-import { ContactRequest, ContactRequestSchema } from "../../../types/Contact";
+import {
+  ContactRequest,
+  ContactRequestSchema,
+} from "../../../types/ContactRequest";
 
 export const sendContactRequest = async (
   username: string,
